+++ include/cacti_parser.h
@@ -1,3 +1,3 @@
 int get_cacti_results(unsigned int SIZE, unsigned int BLOCKSIZE,
-                      unsigned int ASSOC, float *AccessTime, float *Energy,
-                      float *Area);
+                      unsigned int ASSOC, double *AccessTime, double *Energy,
+                      double *Area);
